src/templates/back/packages/auditing/src/@api/auditing/http-communication/resolvers/auditing-upsert-http-communication.resolver.ts
import { Resolver, Args, Mutation } from '@nestjs/graphql';
import { Auditing, AuditingMeta, Timezone } from '@aurora-ts/core';
import { Auth } from '@aurora/decorators';

// @app
import { AuditingUpsertHttpCommunicationHandler } from '../handlers/auditing-upsert-http-communication.handler';
import { AuditingHttpCommunication, AuditingUpdateHttpCommunicationByIdInput } from '@api/graphql';

@Resolver()
@Auth('auditing.httpCommunication.upsert')
export class AuditingUpsertHttpCommunicationResolver
{
    constructor(
        private readonly handler: AuditingUpsertHttpCommunicationHandler,
    ) {}

    @Mutation('auditingUpsertHttpCommunication')
    async main(
        @Args('payload') payload: AuditingUpdateHttpCommunicationByIdInput,
        @Timezone() timezone?: string,
    ): Promise<AuditingHttpCommunication>
    {
        return await this.handler.main(
            payload,
            timezone,
        );
    }
}